requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
supabase>=2.0.0
httpcore>=1.0.0
httpx>=0.27.0
